clean up node server
